+++ domain/src/main/java/net/boreeas/riotapi/constants/QueueId.java
@@ -25,117 +25,117 @@ public class QueueId {
     /**
      * For summoners rift team builder games
      */
-    public final int SR_5v5_NORMAL_CAP = 61;
+    public static final int SR_5v5_NORMAL_CAP = 61;
     /**
      * For summoners rift normal blind games
      */
-    public final int SR_5v5_NORMAL_BLIND = 2;
+    public static final int SR_5v5_NORMAL_BLIND = 2;
     /**
      * For summoner rift normal draft games
      */
-    public final int SR_5v5_NORMAL_DRAFT = 14;
+    public static final int SR_5v5_NORMAL_DRAFT = 14;
     /**
      * For summoners rift ranked solo and duo queue
      */
-    public final int SR_5v5_RANKED_SOLO = 4;
+    public static final int SR_5v5_RANKED_SOLO = 4;
     /**
      * For summoners rift ranked team games
      */
-    public final int SR_5v5_RANKED_TEAM = 42;
+    public static final int SR_5v5_RANKED_TEAM = 42;
     /**
      * For twisted treeline normal blind games
      */
-    public final int TT_3v3_NORMAL_BLIND = 8;
+    public static final int TT_3v3_NORMAL_BLIND = 8;
     /**
      * For twisted treeline ranked team games
      */
-    public final int TT_3v3_RANKED_TEAM = 41;
+    public static final int TT_3v3_RANKED_TEAM = 41;
     /**
      * For crystal scar normal blind games
      */
-    public final int ODIN_5v5_NORMAL_BLIND = 16;
+    public static final int ODIN_5v5_NORMAL_BLIND = 16;
     /**
      * For crystal scar normal draft games
      */
-    public final int ODIN_5v5_NORMAL_DRAFT = 17;
+    public static final int ODIN_5v5_NORMAL_DRAFT = 17;
     /**
      * For howling abyss normal aram games
      */
-    public final int ARAM_5v5_NORMAL_ARAM= 65;
+    public static final int ARAM_5v5_NORMAL_ARAM= 65;
     /**
      * For summoners rift intro bot games
      */
-    public final int SR_5v5_BOT_INTRO = 31;
+    public static final int SR_5v5_BOT_INTRO = 31;
     /**
      * For summoners rift easy bot games
      */
-    public final int SR_5v5_BOT_EASY = 31;
+    public static final int SR_5v5_BOT_EASY = 31;
     /**
      * For summoners rift medium bot games
      */
-    public final int SR_5v5_BOT_MEDIUM = 31;
+    public static final int SR_5v5_BOT_MEDIUM = 31;
     /**
      * For twisted treeline easy and medium bot games
      */
-    public final int TT_3v3_BOT_EASY_AND_MEDIUM = 52;
+    public static final int TT_3v3_BOT_EASY_AND_MEDIUM = 52;
     /**
      * For crystal scar easy and medium bot games
      */
-    public final int ODIN_5v5_BOT_EASY_AND_MEDIUM = 25;
+    public static final int ODIN_5v5_BOT_EASY_AND_MEDIUM = 25;
     /**
      * For nemesis draft featured games
      */
-    public final int FEATURED_COUNTER_PICK = 310;
+    public static final int FEATURED_COUNTER_PICK = 310;
     /**
      * For one-for-all featured games
      */
-    public final int FEATURED_ONE_FOR_ALL = 70;
+    public static final int FEATURED_ONE_FOR_ALL = 70;
     /**
      * For one-for-all in mirror mode featured games.
      */
-    public final int FEATURED_ONE_FOR_ALL_MIRROR_MODE = 78;
+    public static final int FEATURED_ONE_FOR_ALL_MIRROR_MODE = 78;
     /**
      * For one-versus-one featured games
      */
-    public final int FEATURED_SHOWDOWN_1v1 = 72;
+    public static final int FEATURED_SHOWDOWN_1v1 = 72;
     /**
      * For two-versus-two featured games
      */
-    public final int FEATURED_SHOWDOWN_2v2 = 73;
+    public static final int FEATURED_SHOWDOWN_2v2 = 73;
     /**
      * For hexakill on summoners rift featured games
      */
-    public final int FEATURED_HEXAKILL_SR = 75;
+    public static final int FEATURED_HEXAKILL_SR = 75;
     /**
      * For hexakill on twisted treeline featured games
      */
-    public final int FEATURED_HEXAKILL_TT = 98;
+    public static final int FEATURED_HEXAKILL_TT = 98;
     /**
      * For urf featured games
      */
-    public final int FEATURED_URF = 76;
+    public static final int FEATURED_URF = 76;
     /**
      * For ascension featured games
      */
-    public final int FEATURED_ASCENSION = 96;
+    public static final int FEATURED_ASCENSION = 96;
     /**
      * For poroking featured games
      */
-    public final int FEATURED_POROKING = 300;
+    public static final int FEATURED_POROKING = 300;
     /**
      * For stage one doom bots
      */
-    public final int FEATURED_DOOM_BOTS_STAGE_1 = 91;
+    public static final int FEATURED_DOOM_BOTS_STAGE_1 = 91;
     /**
      * For stage two doom bots
      */
-    public final int FEATURED_DOOM_BOTS_STAGE_2 = 92;
+    public static final int FEATURED_DOOM_BOTS_STAGE_2 = 92;
     /**
      * For stage five doom bots
      */
-    public final int FEATURED_DOOM_BOTS_STAGE_5 = 93;
+    public static final int FEATURED_DOOM_BOTS_STAGE_5 = 93;
     /**
      * For URF versus bots
      */
-    public final int FEATURED_URF_BOTS = 83;
+    public static final int FEATURED_URF_BOTS = 83;
 }