+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added missing game modes
 * Added queue type and queue id for the new custom game mode, nemesis pick
 * Added "Master" league tier
+* Added new map (SR Beta)
 
 ##### Rtmp
 * Fixed bug where bot game invites would not pass the queue id